+++ src/features/user/friends/index.ts
@@ -2,8 +2,9 @@
 export { FriendsPanel } from "./components/FriendsPanel";
 
 // Hooks
-export { useUserFriendCount } from "./hooks/useUserFriendCount";
 export { useFriendProfiles } from "./hooks/useFriendProfiles";
+export { useFriendshipStatus } from "./hooks/useFriendshipStatus";
+export { useUserFriendCount } from "./hooks/useUserFriendCount";
 export { useUserFriends } from "./hooks/useUserFriends";
 
 // Types